+++ tests/ParserTest.php
@@ -1,6 +1,7 @@
 <?php
 namespace Thunder\Shortcode\Tests;
 
+use Thunder\Shortcode\HandlerContainer\HandlerContainer;
 use Thunder\Shortcode\Parser\RegularParser;
 use Thunder\Shortcode\Parser\ParserInterface;
 use Thunder\Shortcode\Parser\RegexParser;
@@ -222,6 +223,27 @@ final class ParserTest extends \PHPUnit_Framework_TestCase
         $this->testParser($parser, '[code "xxx"]', array(
             new ParsedShortcode(new Shortcode('code', array('xxx' => null), null, null), '[code "xxx"]', 0)
         ));
+
+        $handlers = new HandlerContainer();
+        $handlers->add('_', function() {});
+        $handlers->add('na_me', function() {});
+        $handlers->add('_n_', function() {});
+        $this->testParser(WordpressParser::createFromHandlers($handlers), '[_][na_me][_name][name_][n_am_e][_n_]', array(
+            new ParsedShortcode(new Shortcode('_', array(), null), '[_]', 0),
+            new ParsedShortcode(new Shortcode('na_me', array(), null), '[na_me]', 3),
+            new ParsedShortcode(new Shortcode('_n_', array(), null), '[_n_]', 32),
+        ));
+        $this->testParser(WordpressParser::createFromNames(array('_', 'na_me', '_n_')), '[_][na_me][_name][name_][n_am_e][_n_]', array(
+            new ParsedShortcode(new Shortcode('_', array(), null), '[_]', 0),
+            new ParsedShortcode(new Shortcode('na_me', array(), null), '[na_me]', 3),
+            new ParsedShortcode(new Shortcode('_n_', array(), null), '[_n_]', 32),
+        ));
+    }
+
+    public function testWordpressInvalidNamesException()
+    {
+        $this->setExpectedException('InvalidArgumentException');
+        WordpressParser::createFromNames(array('string', new \stdClass()));
     }
 
     public function testInstances()